+++ app/src/main/java/com/nutomic/syncthingandroid/onboarding/pages/WelcomePage.kt
@@ -1,0 +1,29 @@
+package com.nutomic.syncthingandroid.onboarding.pages
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import com.nutomic.syncthingandroid.R
+import com.nutomic.syncthingandroid.onboarding.OnboardingIcon
+import com.nutomic.syncthingandroid.onboarding.OnboardingScaffold
+import com.nutomic.syncthingandroid.onboarding.OnboardingUiState
+
+@Composable
+fun WelcomePage(
+    uiState: OnboardingUiState,
+    pageIndex: Int,
+    onBack: () -> Unit,
+    onContinue: () -> Unit,
+) {
+    OnboardingScaffold(
+        icon = OnboardingIcon.Logo,
+        title = stringResource(R.string.welcome_title),
+        description = stringResource(R.string.welcome_text),
+        pageIndex = pageIndex,
+        pageCount = uiState.pages.size,
+        canGoBack = false,
+        backVisible = false,
+        nextLabel = stringResource(R.string.cont),
+        onBack = onBack,
+        onNext = onContinue,
+    )
+}
