+++ app/src/main/java/com/nutomic/syncthingandroid/util/CompositionLocals.kt
@@ -1,13 +1,8 @@
 package com.nutomic.syncthingandroid.util
 
-import androidx.compose.foundation.pager.PagerState
 import androidx.compose.runtime.staticCompositionLocalOf
 import kotlinx.coroutines.CoroutineScope
 
 val LocalActivityScope = staticCompositionLocalOf<CoroutineScope> {
     error("No activity scope provided")
-}
-
-val LocalPagerState = staticCompositionLocalOf<PagerState> {
-    error("No pager state provided")
 }
